move existing exceptions tests to ExceptionTest class
