+++ src/test/java/io/pmem/pmemkv/tests/ExceptionTest.java
@@ -52,6 +52,100 @@ public class ExceptionTest {
         db.stop();
     }
 
+    /* Exceptions related to config and in Open method */
+
+    @Test
+    public void throwsExceptionOnStartWhenPathIsMissing() {
+        Database<ByteBuffer, ByteBuffer> db = null;
+        boolean exception_occured = false;
+        try {
+            db = new Database.Builder<ByteBuffer, ByteBuffer>(ENGINE).
+                setSize(1073741824).
+                build();
+            Assert.fail();
+        } catch (DatabaseException kve) {
+            exception_occured = true;
+        } catch (Exception e) {
+            Assert.fail();
+        }
+        expect(exception_occured).toBeTrue();
+        expect(db).toBeNull();
+    }
+
+    @Test
+    public void throwsExceptionOnStartWhenSizeIsMissing() {
+        Database<ByteBuffer, ByteBuffer> db = null;
+        boolean exception_occured = false;
+        try {
+            db = new Database.Builder<ByteBuffer, ByteBuffer>(ENGINE).
+                setPath("/dev/shm").
+                build();
+            Assert.fail();
+        } catch (DatabaseException kve) {
+            exception_occured = true;
+        } catch (Exception e) {
+            Assert.fail();
+        }
+        expect(db).toBeNull();
+        expect(exception_occured).toBeTrue();
+    }
+
+    @Test
+    public void throwsExceptionOnStartWhenEngineIsInvalidTest() {
+        Database<ByteBuffer, ByteBuffer> db = null;
+        boolean exception_occured = false;
+        try {
+            db = buildDB("nope.nope");
+            Assert.fail();
+        } catch (DatabaseException kve) {
+            exception_occured = true;
+        } catch (Exception e) {
+            Assert.fail();
+        }
+        expect(db).toBeNull();
+        expect(exception_occured).toBeTrue();
+    }
+
+    @Test
+    public void throwsExceptionOnStartWhenPathIsInvalidTest() {
+        Database<ByteBuffer, ByteBuffer> db = null;
+        boolean exception_occured = false;
+        try {
+            db = new Database.Builder<ByteBuffer, ByteBuffer>(ENGINE).
+                setSize(1073741824).
+                setPath("/tmp/123/234/345/456/567/678/nope.nope").
+                build();
+            Assert.fail();
+        } catch (DatabaseException kve) {
+            exception_occured = true;
+        } catch (Exception e) {
+            Assert.fail();
+        }
+        expect(db).toBeNull();
+        expect(exception_occured).toBeTrue();
+    }
+
+    @Test
+    public void throwsExceptionOnStartWhenPathIsWrongTypeTest() {
+        Database<ByteBuffer, ByteBuffer> db = null;
+        boolean exception_occured = false;
+        try {
+            db = new Database.Builder<ByteBuffer, ByteBuffer>(ENGINE).
+                setSize(1073741824).
+                setPath("1234").
+                build();
+            Assert.fail();
+        } catch (DatabaseException kve) {
+            exception_occured = true;
+        } catch (Exception e) {
+            Assert.fail();
+        }
+        expect(exception_occured).toBeTrue();
+        expect(db).toBeNull();
+    }
+
+    /* Exceptions in Gets methods */
+
     @Test
     public void exceptionInGetallTest() {
         int exception_counter = 0;
